requirements.txt
discord.py
aiosqlite
unidecode
jishaku
